requirements.txt
# Django y Base de Datos
Django==4.2.16
psycopg2-binary==2.9.9
python-dotenv==1.0.1
dj-database-url==2.2.0

# Django REST Framework
djangorestframework==3.15.2
djangorestframework-gis==1.0

# CORS
django-cors-headers==4.4.0

# Archivos estáticos
whitenoise==6.7.0

# Imágenes
Pillow==11.0.0

# Utilidades
python-decouple==3.8

# Servidor WSGI para producción
gunicorn==23.0.0

# Desarrollo y Testing (opcionales)
# django-debug-toolbar==4.2.0
# pytest==7.4.3
# pytest-django==4.7.0
# coverage==7.3.2

